# Conflicts:
#	Assembly-CSharp-Editor.csproj
#	Assembly-CSharp.csproj
#	Assets/Code/Game/ILRuntime/Binding/Analysis/BDFramework_ResourceMgr_IResMgr_Binding.cs.meta
#	Assets/Code/Game/ILRuntime/Binding/Analysis/CLRBindings.cs
#	Assets/Code/Game/ILRuntime/Binding/Analysis/SQLite4Unity3d_SQLiteConnection_Binding.cs
#	Assets/StreamingAssets/Android/hotfix/hotfix.dll
#	Assets/StreamingAssets/Android_changelist.json
#	Assets/StreamingAssets/Android_changelist.json.meta
#	Assets/StreamingAssets/StandaloneWindows_changelist.json
#	Assets/StreamingAssets/Windows/Art.meta
#	Assets/StreamingAssets/Windows/Hotfix/hotfix.dll
#	Assets/StreamingAssets/Windows/Hotfix/hotfix.dll.pdb
#	Assets/StreamingAssets/Windows/Local.db
#	Assets/StreamingAssets/iOS/Hotfix/hotfix.dll
#	Packages/manifest.json
#	ProjectSettings/EditorSettings.asset
#	ProjectSettings/ProjectVersion.txt
#	Sirenix.OdinInspector.CompatibilityLayer.Editor.csproj
#	Sirenix.OdinInspector.CompatibilityLayer.csproj

+++ Assets/Code/BDFramework/Editor/ILruntime/EditorWindow_ScriptBuildDll.cs
@@ -101,7 +101,10 @@ public class EditorWindow_ScriptBuildDll : EditorWindow
     {
         var targetPath = "Assets/Code/Game/ILRuntime/Binding/Analysis";
         //分析之前先删除,然后生成临时文件防止报错
-        Directory.Delete(targetPath, true);
+        if (Directory.Exists(targetPath))
+        {
+            Directory.Delete(targetPath, true);
+        }
         var fileContent = @"
 namespace ILRuntime.Runtime.Generated
 {
